# api/openapi.yaml
openapi: 3.0.0
info:
  title: Universal Registrar Driver
  description: Universal Registrar Driver
  contact:
    email: markus@danubetech.com
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/universal-registrar/universal-registrar-driver/1.0.0
  description: SwaggerHub API Auto Mocking
- url: /1.0/
paths:
  /register:
    post:
      summary: Registers a DID.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "200":
          description: successfully registered!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterState'
        "400":
          description: invalid input!
        "500":
          description: error!
      x-swagger-router-controller: DidController
  /update:
    post:
      summary: Updates a DID.
      operationId: update
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRequest'
      responses:
        "200":
          description: successfully updated!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateState'
        "400":
          description: invalid input!
        "500":
          description: error!
      x-swagger-router-controller: DidController
  /deactivate:
    post:
      summary: Deactivates a DID.
      operationId: deactivate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeactivateRequest'
      responses:
        "200":
          description: successfully deactivated!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeactivateState'
        "400":
          description: invalid input!
        "500":
          description: error!
      x-swagger-router-controller: DidController
components:
  schemas:
    RegisterRequest:
      type: object
      properties:
        jobId:
          type: string
          example: 6d85bcd0-2ea3-4288-ab00-15afadd8a156
        options:
          type: object
          example:
            chain: testnet
        secret:
          type: object
          example:
            seed: 72WGp7NgFR1Oqdi8zlt7jQQ434XR0cNQ
        didDocument:
          type: object
      additionalProperties: false
      description: Request parameters for the register operation.
    UpdateRequest:
      required:
      - identifier
      type: object
      properties:
        jobId:
          type: string
          example: 6d85bcd0-2ea3-4288-ab00-15afadd8a156
        identifier:
          type: string
          example: did:sov:WRfXPg8dantKVubE3HX8pw
        options:
          type: object
        secret:
          type: object
        didDocument:
          type: object
      additionalProperties: false
      description: Request parameters for the update operation.
    DeactivateRequest:
      required:
      - identifier
      type: object
      properties:
        jobId:
          type: string
          example: 6d85bcd0-2ea3-4288-ab00-15afadd8a156
        identifier:
          type: string
          example: did:sov:WRfXPg8dantKVubE3HX8pw
        options:
          type: object
        secret:
          type: object
      additionalProperties: false
      description: Request parameters for the deactivate operation.
    RegisterState:
      required:
      - didState
      type: object
      properties:
        jobId:
          type: string
          example: 6d85bcd0-2ea3-4288-ab00-15afadd8a156
        didState:
          $ref: '#/components/schemas/RegisterState_didState'
        registrarMetadata:
          type: object
        methodMetadata:
          type: object
      additionalProperties: false
      description: The state after a register operation.
      example:
        jobId: 6d85bcd0-2ea3-4288-ab00-15afadd8a156
        didState:
          identifier: did:sov:WRfXPg8dantKVubE3HX8pw
          state: finished
          secret:
            seed: 72WGp7NgFR1Oqdi8zlt7jQQ434XR0cNQ
            keys:
            - privateKeyJwk:
                kty: OKP
                d: NzJXR3A3TmdGUjFPcWRpOHpsdDdqUVE0MzRYUjBjTlE
                crv: Ed25519
                x: jpIKKU2b77lNXKTNW2NGvw1GUMjU6v_l_tLJAH5uYz0
              publicKeyDIDURL: did:sov:WRfXPg8dantKVubE3HX8pw#key-1
        methodMetadata: {}
        registrarMetadata: {}
    UpdateState:
      required:
      - didState
      type: object
      properties:
        jobId:
          type: string
          example: 6d85bcd0-2ea3-4288-ab00-15afadd8a156
        didState:
          $ref: '#/components/schemas/UpdateState_didState'
        registrarMetadata:
          type: object
        methodMetadata:
          type: object
      additionalProperties: false
      description: The state after an update operation.
      example:
        jobId: 6d85bcd0-2ea3-4288-ab00-15afadd8a156
        didState:
          state: state
          secret: {}
        methodMetadata: {}
        registrarMetadata: {}
    DeactivateState:
      required:
      - didState
      type: object
      properties:
        jobId:
          type: string
          example: 6d85bcd0-2ea3-4288-ab00-15afadd8a156
        didState:
          $ref: '#/components/schemas/DeactivateState_didState'
        registrarMetadata:
          type: object
        methodMetadata:
          type: object
      additionalProperties: false
      description: The state after a deactivate operation.
      example:
        jobId: 6d85bcd0-2ea3-4288-ab00-15afadd8a156
        didState:
          state: state
        methodMetadata: {}
        registrarMetadata: {}
    RegisterState_didState:
      type: object
      properties:
        state:
          type: string
          example: finished
        identifier:
          type: string
          example: did:sov:WRfXPg8dantKVubE3HX8pw
        secret:
          type: object
          example:
            seed: 72WGp7NgFR1Oqdi8zlt7jQQ434XR0cNQ
            keys:
            - privateKeyJwk:
                kty: OKP
                d: NzJXR3A3TmdGUjFPcWRpOHpsdDdqUVE0MzRYUjBjTlE
                crv: Ed25519
                x: jpIKKU2b77lNXKTNW2NGvw1GUMjU6v_l_tLJAH5uYz0
              publicKeyDIDURL: did:sov:WRfXPg8dantKVubE3HX8pw#key-1
      example:
        identifier: did:sov:WRfXPg8dantKVubE3HX8pw
        state: finished
        secret:
          seed: 72WGp7NgFR1Oqdi8zlt7jQQ434XR0cNQ
          keys:
          - privateKeyJwk:
              kty: OKP
              d: NzJXR3A3TmdGUjFPcWRpOHpsdDdqUVE0MzRYUjBjTlE
              crv: Ed25519
              x: jpIKKU2b77lNXKTNW2NGvw1GUMjU6v_l_tLJAH5uYz0
            publicKeyDIDURL: did:sov:WRfXPg8dantKVubE3HX8pw#key-1
    UpdateState_didState:
      type: object
      properties:
        state:
          type: string
        secret:
          type: object
      example:
        state: state
        secret: {}
    DeactivateState_didState:
      type: object
      properties:
        state:
          type: string
      example:
        state: state